requirements.txt
# Core Django
Django==4.2.18
djangorestframework==3.14.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching and Background Tasks
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Environment and Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Security and Monitoring
django-ratelimit==4.1.0
sentry-sdk==1.40.6

# Static Files and Media
whitenoise==6.7.0
pillow==10.4.0
brotli==1.1.0

# Web Scraping and AI
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.6.1
yt-dlp==2023.12.30

# Audio Processing (for YouTube module)
whisper==1.1.10
torch==2.1.2
torchaudio==2.1.2

# Scheduling
schedule==1.2.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
coverage==7.3.2

# Production Server
gunicorn==21.2.0
uvicorn==0.24.0

# Utilities
requests==2.31.0
python-slugify==8.0.1
Faker==20.1.0
